# Conflicts:
#	src/common/common-int-test/src/main/resources/signer-container-files/Dockerfile

+++ Docker/securityserver/init_context.sh
@@ -1,6 +1,6 @@
 #!/bin/bash
 
-rm -rf ./build
+rm -rf ./build/packages
 mkdir -p ./build/packages
 mkdir -p ./build/libs
 
@@ -13,5 +13,9 @@ cp -r ../../src/security-server/edc/runtime/data-plane/src/main/resources/liquib
 cp -r ../../src/security-server/edc/runtime/identity-hub/src/main/resources/liquibase ./build
 
 # Should match opentelemetry annotations version.
-wget -O ./build/libs/opentelemetry-javaagent.jar \
-https://github.com/open-telemetry/opentelemetry-java-instrumentation/releases/download/v2.6.0/opentelemetry-javaagent.jar
+if [ ! -f ./build/libs/opentelemetry-javaagent.jar ]; then
+  wget -O ./build/libs/opentelemetry-javaagent.jar \
+  https://github.com/open-telemetry/opentelemetry-java-instrumentation/releases/download/v2.9.0/opentelemetry-javaagent.jar
+else
+  echo "opentelemetry-javaagent.jar already exists. Skipping download."
+fi
